Add cache busting code to simplify releases (#44)
When pushed to production, all network traffic goes through Cloudflare,
where they cache the content of the website. In order for releases to
go smoothly, we want to add a nonce at the end of the request that is
unique to each release. Because its stable, the cdn can use its cache
if it sees the same nonce. However, each release will introduce a new
key, which ensures new content is served on each release.

* Updated index.html to support the extra query param for cache busting
* Updated Dockerfile to compile new index.html before passing it to
nginx
* Update app.py for new template context requirement
* Update gitignore for files generated by VIM
* Add new release image building script
* Clean up outdated prod directory